feat: v1.4.0 — 14 bug fixes, pixi.toml, Python 3.13, install UX (#68)
* fix: resolve 14 verified bugs across Python and Rust pipelines

Python fixes:
- fix(count_alleles): use ~ instead of not for pandas Series boolean negation (#50)
- fix(count_alleles_sc): uint8 overflow → uint16 for SC allele counts (#60)
- fix(run_counting_sc): init is_gene_file=False when no feature file (#53)
- fix(parse_gene_data): .all() → .any() for GTF attribute detection (#64)
- fix(counting/__main__): pass full sample list instead of samples[0] (#51)
- fix(mapping/__main__): pass full sample list instead of samples[0] (#51)
- fix(analysis/__main__): pass full groups list instead of groups[0] (#51)
- fix(run_analysis): warn when --phased/--region_col/--groupby ignored (#52)
- fix(run_analysis_sc): issubset for phased genotype detection (#63)
- fix(compare_ai): dead expression → assign snp_cutoff correctly (#55)
- fix(run_mapping): catch NameError on json_dict in decorator (#54)
- fix(wasp_data_files): sample 100 VCF records for phase detection (#65)

Rust fixes:
- fix(analysis): sort par_iter results for deterministic output (#62)
- fix(mapping_filter): separate removed_moved from missing_count stats (#58)
- fix(unified_pipeline): check remap_names_path in parallel fallback (#57)
- fix(vcf_to_bed): BGZF fallback to standard gzip via flate2 (#59)
- fix(vcf_to_bed): document noodles Debug format usage for genotypes (#56)


* fix: resolve linear model shape mismatch and test_opt_prob missing arg

- Fix opt_unphased_dp to split per-variant dispersion array for first vs
  subsequent positions when using the linear model. The disp array (size N)
  was being broadcast against phase_ref/phase_n arrays (size N-1), causing
  ValueError in scipy's betabinom.
- Fix test_opt_prob_large: add missing in_rho argument to opt_prob call
  and update assertion to accept array return type.

All 139 tests now pass (previously 5 failures).


* feat: add pixi.toml and improve install UX for deployment

- Add pixi.toml for one-command install of all dependencies (Python,
  Rust, samtools, bcftools, bedtools, htslib) via pixi
- Update README and installation docs to recommend pixi first
- Fix bioconda recipe: remove conflicting script directive so build.sh
  is used, add --locked flag per bioconda policy
- Remove defaults channel from environment.yml (dropped by bioconda)
- Add Python 3.13 and --locked to release.yml wheel builds
- Update wheel count validation from 12 to 16


* chore: add Python 3.13 classifier to pyproject.toml

Aligns PyPI metadata with the release.yml wheel build matrix
which now builds for Python 3.10-3.13.


* docs: recommend mamba over conda for installation

Mamba uses the same bioconda packages but has a much faster dependency
solver. Updated installation docs and README to recommend mamba first,
with conda as fallback.


* chore: bump version to 1.4.0 for release

Version bump across all 6 packaging files:
- rust/Cargo.toml (source of truth)
- rust/Cargo.lock
- pixi.toml (also widen Python to <3.14 for 3.13 support)
- bioconda-recipe/meta.yaml
- Dockerfile
- Singularity.def

Verified with scripts/check-version-consistency.sh — all consistent.


---------

+++ src/counting/count_alleles_sc.py
@@ -138,19 +138,19 @@ def make_count_matrix(
     sparse_ref = csr_matrix(
         (list(sc_counts.ref_count.values()), list(zip(*sc_counts.ref_count.keys()))),
         shape=(snp_df.shape[0], len(bc_dict)),
-        dtype=np.uint8,
+        dtype=np.uint16,
     )
 
     sparse_alt = csr_matrix(
         (list(sc_counts.alt_count.values()), list(zip(*sc_counts.alt_count.keys()))),
         shape=(snp_df.shape[0], len(bc_dict)),
-        dtype=np.uint8,
+        dtype=np.uint16,
     )
 
     sparse_other = csr_matrix(
         (list(sc_counts.other_count.values()), list(zip(*sc_counts.other_count.keys()))),
         shape=(snp_df.shape[0], len(bc_dict)),
-        dtype=np.uint8,
+        dtype=np.uint16,
     )
 
     # Create anndata With total as X
